Moved TestSession into ekat_test_utils.hpp
This way it can be accessed by a test source, and set particular parameters.
